Fix DELREQUIRES
When DELREQUIRES is specified, the old algorithm's double for-loop did
not actually remove any of the DELREQUIRES entries from the REQUIRES
list.  Switch the logic of the double for-loop around so that every item
in INFOREQUIRES is tested against the entire list of DELREQUIRES (and
'%README%'), and only when it never matches then add it to newreqlist.

The "break" statement in the inner for-loop is an optimization and not
strictly needed.  It bails out of the inner for-loop as soon as we've
found a match in the DELREQUIRES (or '%README%') list.